requirements.txt
streamlit==1.43.0
requests==2.32.3
python-dotenv==1.0.1